LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-self-alignment-positioned-items-with-margin-border-padding-016-expected.txt

X XX XXX X
X XXX X
XX XXXX XX XXX X
X XXX
X
XX XXX X XX XXX X
X XXX X
XX XXXX XX XXX X
X XXX
X
XX XXX

X XX XXX X
X XXX X
XX XXXX XX XXX X
X XXX
X
XX XXX X XX XXX X
X XXX X
XX XXXX XX XXX X
X XXX
X
XX XXX

PASS .grid 1
FAIL .grid 2 assert_equals:
<div class="grid RTL">
  <div data-offset-x="377" data-offset-y="3" data-expected-width="76" data-expected-height="20" class="firstRowFirstColumn">X XX X</div>
  <div data-offset-x="86" data-offset-y="105" data-expected-width="86" data-expected-height="40" class="firstRowSecondColumn">XX X<br>X XXX X<br>XX XXX</div>
  <div data-offset-x="377" data-offset-y="189" data-expected-width="76" data-expected-height="20" class="secondRowFirstColumn">X XX X</div>
  <div data-offset-x="261" data-offset-y="195" data-expected-width="76" data-expected-height="50" class="secondRowSecondColumn">XX X<br>X XXX<br>X<br>XX XXX</div>
</div>
offsetLeft expected 377 but got 375
PASS .grid 3
PASS .grid 4
